Updated requests processing code

+++ py-modules/tileserver-stats/tileserver_stats/__init__.py
@@ -7,6 +7,7 @@ from typer import Typer
 from dotenv import load_dotenv
 from sqlalchemy.sql import text
 from datetime import datetime
+from psycopg2.sql import Literal
 
 load_dotenv()
 
@@ -17,12 +18,14 @@ app.command()
 
 @app.command()
 def run():
-
+    """Run the update procedure."""
     tileserver_db = environ.get("TILESERVER_STATS_DATABASE")
     db = Database(tileserver_db)
 
     # Run update
     fn = Path(relative_path(__file__, "procedures")) / "run-update.sql"
+    sql = text(fn.read_text().replace(":", "\:"))
+
 
     # check query timing
     conn = db.engine.connect()
@@ -30,9 +33,9 @@ def run():
     start = datetime.now()
     step = start
     while n_results > 0:
-        res = conn.execute(fn.read_text()).first()
+        res = conn.execute(sql, execution_options=dict(no_parameters=True)).first()
         n_results = res.n_rows
-        conn.execute("COMMIT")
+        conn.execute(text("COMMIT"))
         next_step = datetime.now()
         dt = (next_step - step).total_seconds()
         print(f"{res.last_row_id} ({dt*1000:.0f} ms)")
@@ -42,10 +45,13 @@ def run():
 
 
 @app.command()
-def reset():
+def reset(drop: bool = False):
+    """Empty the stats schema and re-run the schema creation scripts."""
     tileserver_db = environ.get("TILESERVER_STATS_DATABASE")
     db = Database(tileserver_db)
-    db.engine.execute("DROP SCHEMA IF EXISTS stats CASCADE")
+
+    if drop:
+        db.engine.execute("DROP SCHEMA IF EXISTS stats CASCADE")
 
     files = Path(relative_path(__file__, "schema")).glob("*.sql")
     files = list(files)
